requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
reportlab==4.2.2